+++ core/runtime/piper/js/js_context_wrapper.h
@@ -12,7 +12,10 @@
 
 #include "core/base/lynx_export.h"
 #include "core/runtime/bindings/jsi/global.h"
+#include "core/runtime/bindings/napi/napi_environment.h"
 #include "core/runtime/jsi/jsi.h"
+#include "core/runtime/piper/js/runtime_lifecycle_listener_delegate.h"
+#include "core/runtime/piper/js/runtime_lifecycle_observer_impl.h"
 #include "core/runtime/profile/runtime_profiler.h"
 
 namespace lynx {
@@ -33,10 +36,12 @@ class LYNX_EXPORT_FOR_DEVTOOL JSContextWrapper
       std::shared_ptr<piper::Runtime>& js_runtime,
       std::shared_ptr<piper::ConsoleMessagePostMan> post_man,
       const tasm::PageOptions& page_options) = 0;
+  virtual void AddLifecycleListener(
+      std::unique_ptr<RuntimeLifecycleListenerDelegate> listener){};
 
   bool isGlobalInited() { return global_inited_; }
   bool isJSCoreLoaded() { return js_core_loaded_; }
-  void loadPreJS(
+  void prepareJSEnv(
       std::weak_ptr<piper::Runtime> js_runtime,
       std::vector<std::pair<std::string, std::shared_ptr<piper::Buffer>>>&
           js_preload);
@@ -48,6 +53,7 @@ class LYNX_EXPORT_FOR_DEVTOOL JSContextWrapper
       std::shared_ptr<profile::RuntimeProfiler> runtime_profiler);
 #endif
  protected:
+  virtual void InitNapi(std::shared_ptr<piper::Runtime>& js_runtime){};
   std::weak_ptr<piper::JSIContext> js_context_;
   bool js_core_loaded_;
   bool global_inited_;
@@ -66,7 +72,7 @@ class LYNX_EXPORT_FOR_DEVTOOL SharedJSContextWrapper : public JSContextWrapper {
   SharedJSContextWrapper(std::shared_ptr<piper::JSIContext>,
                          const std::string& group_id,
                          ReleaseListener* listener);
-  ~SharedJSContextWrapper() = default;
+  ~SharedJSContextWrapper() override = default;
 
   virtual void Def() override;
   virtual void EnsureConsole(
@@ -77,10 +83,18 @@ class LYNX_EXPORT_FOR_DEVTOOL SharedJSContextWrapper : public JSContextWrapper {
                   std::shared_ptr<piper::ConsoleMessagePostMan> post_man,
                   const tasm::PageOptions& page_options) override;
 
+  void AddLifecycleListener(
+      std::unique_ptr<RuntimeLifecycleListenerDelegate> listener) override;
+
  protected:
+  void InitNapi(std::shared_ptr<piper::Runtime>& js_runtime) override;
   std::shared_ptr<piper::SharedContextGlobal> global_;
   std::string group_id_;
   ReleaseListener* listener_;
+#if ENABLE_NAPI_BINDING
+  std::unique_ptr<piper::NapiEnvironment> napi_environment_;
+  std::unique_ptr<RuntimeLifecycleObserverImpl> lifecycle_observer_;
+#endif
 };
 
 class LYNX_EXPORT_FOR_DEVTOOL NoneSharedJSContextWrapper